docs: type alias added.
I do not really understand why this is necessary - logic would tell you replacing something by the exact thing should not change anything. But the fact is that there are three options for the flow_operations docs:

- use type "Flow" (as imported). Renders as "oflibnumpy.Flow" in the docs, but is not hyperlinked (why, when it clearly understands which class it is and where it comes from?)
- use type "FlowAlias" defined at the start of the file (FlowAlias = 'Flow'). Renders as "oflibnumpy.flow_class.Flow" in the docs, but is also not hyperlinked (this time, it appears to look at the file path provenance rather than just the code hierarchy)
- use type "FlowAlias" as before, *and* use "autodoc_type_aliases = {"Flow": "Flow"}" in conf.py. This renders as "Flow" in the docs (as desired), and is now also hyperlinked to the oflibnumpy.Flow doc.

I suppose it has to do with replacing the literal string "Flow" with the sphinx-internal understanding of "Flow", which is actually linked to the known Flow class and therefore renders with the hyperlink.